+++ frontend/src/types/tls.ts
@@ -1,5 +1,12 @@
 import { Dial } from "./dial"
 
+export interface tls {
+  id: number
+  name: string
+  server: iTls
+  client: oTls
+}
+
 export interface iTls {
   enabled?: boolean
   server_name?: string
@@ -62,4 +69,55 @@ export const defaultInTls: iTls = {
   min_version: "1.2",
   max_version: "1.3",
   cipher_suites: [],
+}
+
+export interface oTls {
+  enabled?: boolean
+  disable_sni?: boolean
+  server_name?: string
+  insecure?: boolean
+  alpn?: string[]
+  min_version?: string
+  max_version?: string
+  cipher_suites?: string[]
+  certificate?: string
+  certificate_path?: string
+  ech?: {
+    enabled: boolean
+    pq_signature_schemes_enabled?: boolean
+    dynamic_record_sizing_disabled?: boolean
+    config?: string[],
+    config_path?: string
+  },
+  utls?: {
+    enabled: boolean
+    fingerprint: string
+  },
+  reality?: {
+    enabled: boolean
+    public_key: string
+    short_id: string
+  }
+}
+
+export const defaultOutTls: oTls = {
+  alpn: ['h3', 'h2', 'http/1.1'],
+  min_version: "1.2",
+  max_version: "1.3",
+  cipher_suites: [],
+  utls: {
+    enabled: true,
+    fingerprint: "chrome",
+  },
+  reality: {
+    enabled: true,
+    public_key: "",
+    short_id: "",
+  },
+  ech: {
+    enabled: true,
+    pq_signature_schemes_enabled: false,
+    dynamic_record_sizing_disabled: false,
+    config_path: "",
+  }
 }
